assetcache/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-py==0.6.0
alpaca-trade-api==2.3.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
deprecation==2.1.0
falcon==3.1.0
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
msgpack==1.0.3
multidict==6.0.2
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sseclient-py==1.7.2
tinydb==4.7.0
typing-extensions==4.3.0
urllib3==1.26.12
websocket-client==1.4.1
websockets==10.3
yarl==1.8.1
